cli/myci: Learn how to talk to https server
We need access to jenkins_cert.pem to be able to talk to the selfsigned
certificate. Of course if you have a proper signed certificate this
won't be necessary. But since this server is more likely a homebrew
version, let's allow minimum hassle setup that is still somewhat secure.

New config value ssl_ca is added to store the location of
jenkins_cert.pem.

jenkins_cert.pem will be generated automatically when running
setup_jenkins.sh

Signed-off-by: Qais Yousef <qyousef@layalina.io>