Yet another fix for jenkins_home permission issue
docker run -u $(id -u) is not good enough. The gid is set to root still
which makes it insecure.

In this, hopefully last, attempt to address this, we move to using
volumes to manage jenkins_home. This means we need to copy things in and
out of the volume but hopefully with the given scripts that extra step
shouldn't prove to be too much of a hassle.

Needs more testing indeed but initial results looked fine. Fingers
crossed!

Signed-off-by: Qais Yousef <qyousef@layalina.io>