Use jenkins.war instead of docker
Using docker has proved to be very cumbersome in terms of easily develop
new tests (keeping files in sync in and out of the volume), beside the
multitude of workarounds to handle the permission issues.

Simplify the process by using jenkins.war. Major issues is handling Java
versions which was hidden by the docker image. Hopefully this won't
prove to be a problem in the long run.

It should be easier now to handle java_home to add new submodules and
develop new tests.

Signed-off-by: Qais Yousef <qyousef@layalina.io>