disconnect: Kill adb server daemon of no devices are attached
This forces a fresh server to start on next connection. There are weird
adb connection stability issues that might be attributed to long running
server daemon going funny. This is an attempt to improve the stability
by regularly killing the daemon when no devices are attached so that the
daemon is refreshed on next connection.

Signed-off-by: Qais Yousef <qyousef@layalina.io>